stats: (hopefully) fix 'Unintended sign extension (SIGN_EXTENSION)'
sign_extension: Suspicious implicit sign extension: "pcr[0]"
with type "unsigned char" (8 bits, unsigned) is promoted in
"pcr[0] * 33554432 + pcr[1] * 131072 + pcr[2] * 512 + pcr[3]
 * 2 + ((pcr[4] & 0x80) >> 7)" to type "int" (32 bits, signed),
then sign-extended to type "unsigned long" (64 bits, unsigned).
If "pcr[0] * 33554432 + pcr[1] * 131072 + pcr[2] * 512 + pcr[3]
 * 2 + ((pcr[4] & 0x80) >> 7)" is greater than 0x7FFFFFFF,
the upper bits of the result will all be 1.

+++ libdvbtee/stats.cpp
@@ -67,11 +67,11 @@ stats& stats::operator= (const stats& cSource)
 
 static void parse_pcr(uint8_t *pcr, uint64_t *pcr_base, unsigned int *pcr_ext)
 {
-	*pcr_base = (pcr[0] * 0x2000000) +
-		    (pcr[1] * 0x20000) +
-		    (pcr[2] * 0x200) +
-		    (pcr[3] * 0x02) +
-		    ((pcr[4] & 0x80) >> 7);
+	*pcr_base = ((uint64_t)pcr[0] * 0x2000000) +
+		    ((uint64_t)pcr[1] * 0x20000) +
+		    ((uint64_t)pcr[2] * 0x200) +
+		    ((uint64_t)pcr[3] * 0x02) +
+		    ((uint64_t)(pcr[4] & 0x80) >> 7);
 
 	*pcr_ext = ((pcr[4] & 0x01) * 0x100) + pcr[5];
